Started moving functions from tx.rs and rpc.rs.
