+++ apps/src/lib/client/signing.rs
@@ -16,6 +16,7 @@ use namada::ledger::rpc::TxBroadcastData;
 use namada::ledger::wallet::Wallet;
 use namada::ledger::wallet::WalletUtils;
 use crate::facade::tendermint_rpc::Client;
+use namada::ledger::signing::TxSigningKey;
 
 /// Find the public key for the given address and try to load the keypair
 /// for it from the wallet. Panics if the key cannot be found or loaded.
@@ -24,63 +25,7 @@ pub async fn find_keypair<C: Client + namada::ledger::queries::Client + Sync, U:
     wallet: &mut Wallet<P>,
     addr: &Address,
 ) -> common::SecretKey {
-    match addr {
-        Address::Established(_) => {
-            println!(
-                "Looking-up public key of {} from the ledger...",
-                addr.encode()
-            );
-            let public_key = rpc::get_public_key(client, addr)
-                .await
-                .unwrap_or_else(|| {
-                    eprintln!(
-                        "No public key found for the address {}",
-                        addr.encode()
-                    );
-                    cli::safe_exit(1);
-                });
-            wallet.find_key_by_pk::<U>(&public_key).unwrap_or_else(|err| {
-                eprintln!(
-                    "Unable to load the keypair from the wallet for public \
-                     key {}. Failed with: {}",
-                    public_key, err
-                );
-                cli::safe_exit(1)
-            })
-        }
-        Address::Implicit(ImplicitAddress(pkh)) => {
-            wallet.find_key_by_pkh::<U>(pkh).unwrap_or_else(|err| {
-                eprintln!(
-                    "Unable to load the keypair from the wallet for the \
-                     implicit address {}. Failed with: {}",
-                    addr.encode(),
-                    err
-                );
-                cli::safe_exit(1)
-            })
-        }
-        Address::Internal(_) => {
-            eprintln!(
-                "Internal address {} doesn't have any signing keys.",
-                addr
-            );
-            cli::safe_exit(1)
-        }
-    }
-}
-
-/// Carries types that can be directly/indirectly used to sign a transaction.
-#[allow(clippy::large_enum_variant)]
-#[derive(Clone)]
-pub enum TxSigningKey {
-    // Do not sign any transaction
-    None,
-    // Obtain the actual keypair from wallet and use that to sign
-    WalletKeypair(common::SecretKey),
-    // Obtain the keypair corresponding to given address from wallet and sign
-    WalletAddress(Address),
-    // Directly use the given secret key to sign transactions
-    SecretKey(common::SecretKey),
+    namada::ledger::signing::find_keypair::<C, U, P>(client, wallet, addr).await
 }
 
 /// Given CLI arguments and some defaults, determine the rightful transaction
@@ -93,46 +38,7 @@ pub async fn tx_signer<C: Client + namada::ledger::queries::Client + Sync, U: Wa
     args: &args::Tx,
     mut default: TxSigningKey,
 ) -> common::SecretKey {
-    // Override the default signing key source if possible
-    if let Some(signing_key) = &args.signing_key {
-        default = TxSigningKey::WalletKeypair(signing_key.clone());
-    } else if let Some(signer) = &args.signer {
-        default = TxSigningKey::WalletAddress(signer.clone());
-    }
-    // Now actually fetch the signing key and apply it
-    match default {
-        TxSigningKey::WalletKeypair(signing_key) => {
-            signing_key
-        }
-        TxSigningKey::WalletAddress(signer) => {
-            let signer = signer;
-            let signing_key = find_keypair::<C, U, P>(
-                client,
-                wallet,
-                &signer,
-            )
-            .await;
-            // Check if the signer is implicit account that needs to reveal its
-            // PK first
-            if matches!(signer, Address::Implicit(_)) {
-                let pk: common::PublicKey = signing_key.ref_to();
-                super::tx::reveal_pk_if_needed::<C, U, P>(client, wallet, &pk, args).await;
-            }
-            signing_key
-        }
-        TxSigningKey::SecretKey(signing_key) => {
-            // Check if the signing key needs to reveal its PK first
-            let pk: common::PublicKey = signing_key.ref_to();
-            super::tx::reveal_pk_if_needed::<C, U, P>(client, wallet, &pk, args).await;
-            signing_key
-        }
-        TxSigningKey::None => {
-            panic!(
-                "All transactions must be signed; please either specify the \
-                 key or the address from which to look up the signing key."
-            );
-        }
-    }
+    namada::ledger::signing::tx_signer::<C, U, P>(client, wallet, args, default).await
 }
 
 /// Sign a transaction with a given signing key or public key of a given signer.
@@ -150,17 +56,7 @@ pub async fn sign_tx<C: Client + namada::ledger::queries::Client + Sync, U: Wall
     args: &args::Tx,
     default: TxSigningKey,
 ) -> TxBroadcastData {
-    let keypair = tx_signer::<C, U, P>(client, wallet, args, default).await;
-    let tx = tx.sign(&keypair);
-
-    let epoch = rpc::query_epoch(client)
-    .await;
-    let broadcast_data = if args.dry_run {
-        TxBroadcastData::DryRun(tx)
-    } else {
-        sign_wrapper(args, epoch, tx, &keypair).await
-    };
-    broadcast_data
+    namada::ledger::signing::sign_tx::<C, U, P>(client, wallet, tx, args, default).await
 }
 
 /// Create a wrapper tx from a normal tx. Get the hash of the
@@ -172,31 +68,5 @@ pub async fn sign_wrapper(
     tx: Tx,
     keypair: &common::SecretKey,
 ) -> TxBroadcastData {
-    let tx = {
-        WrapperTx::new(
-            Fee {
-                amount: args.fee_amount,
-                token: args.fee_token.clone(),
-            },
-            keypair,
-            epoch,
-            args.gas_limit.clone(),
-            tx,
-            // TODO: Actually use the fetched encryption key
-            Default::default(),
-        )
-    };
-
-    // We use this to determine when the wrapper tx makes it on-chain
-    let wrapper_hash = hash_tx(&tx.try_to_vec().unwrap()).to_string();
-    // We use this to determine when the decrypted inner tx makes it
-    // on-chain
-    let decrypted_hash = tx.tx_hash.to_string();
-    TxBroadcastData::Wrapper {
-        tx: tx
-            .sign(keypair)
-            .expect("Wrapper tx signing keypair should be correct"),
-        wrapper_hash,
-        decrypted_hash,
-    }
+    namada::ledger::signing::sign_wrapper(args, epoch, tx, keypair).await
 }
