+++ shared/Cargo.toml
@@ -136,6 +136,7 @@ zeroize = "1.5.5"
 toml = "0.5.8"
 bimap = {version = "0.6.2", features = ["serde"]}
 orion = "0.16.0"
+tokio = {version = "1.8.2"}
 
 [dev-dependencies]
 assert_matches = "1.5.0"
