+++ shared/src/ledger/mod.rs
@@ -14,6 +14,8 @@ pub mod vp_host_fns;
 pub mod wallet;
 pub mod args;
 pub mod rpc;
+pub mod tx;
+pub mod signing;
 
 pub use namada_core::ledger::{
     gas, governance, parameters, storage_api, tx_env, vp_env,
